requirements.txt
asgiref==3.11.0
Django==4.2.27
django-filter==25.1
django-jazzmin==3.0.1
djangorestframework==3.16.1
importlib_metadata==8.7.1
Markdown==3.9
pillow==11.3.0
sqlparse==0.5.5
typing_extensions==4.15.0
zipp==3.23.0
